fix: 폴더이름 변경
